Remove _traversal helpers, move dependency graph generation to separate file
